+++ src/idunno.Bluesky/Embed/EmbeddedExternal.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Serialization;
 
 using idunno.AtProto;
@@ -28,27 +29,13 @@ public sealed record EmbeddedExternal : EmbeddedBase
     /// <summary>
     /// Creates a new instance of <see cref="EmbeddedExternal"/>.
     /// </summary>
-    /// <param name="uri">The external <see cref="Uri"/> for the link.</param>
-    /// <param name="title">The title for the external link.</param>
-    /// <param name="description">The description of the external link, if any.</param>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/>, <paramref name="title"/>, or <paramref name="description"/> is <see langword="null"/>.</exception>
-    public EmbeddedExternal(Uri uri, string title, string description) :
-        this(new External.Properties(uri, title, description, thumbnail: null))
-    {
-        ArgumentNullException.ThrowIfNull(uri);
-        ArgumentNullException.ThrowIfNull(title);
-        ArgumentNullException.ThrowIfNull(description);
-    }
-
-    /// <summary>
-    /// Creates a new instance of <see cref="EmbeddedExternal"/>.
-    /// </summary>
-    /// <param name="uri">The external <see cref="Uri"/> for the link.</param>
+    /// <param name="uri">The external uri for the link, as a string.</param>
     /// <param name="title">The title for the external link.</param>
     /// <param name="description">The description of the external link, if any.</param>
     /// <param name="thumbnail">The <see cref="Blob"/> for the thumbnail of the link, if any.</param>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/>, <paramref name="title"/>, or <paramref name="description"/> is <see langword="null"/>.</exception>
-    public EmbeddedExternal(Uri uri, string title, string description, Blob? thumbnail) :
+    [SuppressMessage("Design", "CA1054:URI-like parameters should not be strings", Justification = "The Bluesky web app can create facets with illegal URIs, so a string is used to accommodate them.")]
+    public EmbeddedExternal(string uri, string title, string description, Blob? thumbnail) :
         this(new External.Properties(uri, title, description, thumbnail))
     {
         ArgumentNullException.ThrowIfNull(uri);
@@ -59,62 +46,15 @@ public sealed record EmbeddedExternal : EmbeddedBase
     /// <summary>
     /// Creates a new instance of <see cref="EmbeddedExternal"/>.
     /// </summary>
-    /// <param name="uri">The external <see cref="Uri"/> for the link.</param>
+    /// <param name="uri">The external uri for the link, as a string.</param>
     /// <param name="title">The title for the external link.</param>
     /// <param name="description">The description of the external link, if any.</param>
     /// <param name="thumbnail">The <see cref="Blob"/> for the thumbnail of the link, if any.</param>
     /// <param name="associatedRefs">An array of strong references associated with the embed.</param>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/>, <paramref name="title"/>, or <paramref name="description"/> is <see langword="null"/>.</exception>
-    public EmbeddedExternal(Uri uri, string title, string description, Blob? thumbnail, StrongReference[]? associatedRefs) :
+    [SuppressMessage("Design", "CA1054:URI-like parameters should not be strings", Justification = "The Bluesky web app can create facets with illegal URIs, so a string is used to accommodate them.")]
+    public EmbeddedExternal(string uri, string title, string description, Blob? thumbnail, StrongReference[]? associatedRefs) :
         this(new External.Properties(uri, title, description, thumbnail, associatedRefs))
-    {
-        ArgumentNullException.ThrowIfNull(uri);
-        ArgumentNullException.ThrowIfNull(title);
-        ArgumentNullException.ThrowIfNull(description);
-    }
-
-    /// <summary>
-    /// Creates a new instance of <see cref="EmbeddedExternal"/>.
-    /// </summary>
-    /// <param name="uri">The external <see cref="Uri"/> for the link.</param>
-    /// <param name="title">The title for the external link.</param>
-    /// <param name="description">The description of the external link, if any.</param>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/>, <paramref name="title"/>, or <paramref name="description"/> is <see langword="null"/>.</exception>
-    public EmbeddedExternal(string uri, string title, string description) :
-        this(new External.Properties(new Uri(uri), title, description, thumbnail: null))
-    {
-        ArgumentNullException.ThrowIfNull(uri);
-        ArgumentNullException.ThrowIfNull(title);
-        ArgumentNullException.ThrowIfNull(description);
-    }
-
-    /// <summary>
-    /// Creates a new instance of <see cref="EmbeddedExternal"/>.
-    /// </summary>
-    /// <param name="uri">The external <see cref="Uri"/> for the link.</param>
-    /// <param name="title">The title for the external link.</param>
-    /// <param name="description">The description of the external link, if any.</param>
-    /// <param name="thumbnail">The <see cref="Blob"/> for the thumbnail of the link, if any.</param>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/>, <paramref name="title"/>, or <paramref name="description"/> is <see langword="null"/>.</exception>
-    public EmbeddedExternal(string uri, string title, string description, Blob? thumbnail) :
-        this(new External.Properties(new Uri(uri), title, description, thumbnail))
-    {
-        ArgumentNullException.ThrowIfNull(uri);
-        ArgumentNullException.ThrowIfNull(title);
-        ArgumentNullException.ThrowIfNull(description);
-    }
-
-    /// <summary>
-    /// Creates a new instance of <see cref="EmbeddedExternal"/>.
-    /// </summary>
-    /// <param name="uri">The external <see cref="Uri"/> for the link.</param>
-    /// <param name="title">The title for the external link.</param>
-    /// <param name="description">The description of the external link, if any.</param>
-    /// <param name="thumbnail">The <see cref="Blob"/> for the thumbnail of the link, if any.</param>
-    /// <param name="associatedRefs">An array of strong references associated with the embed.</param>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/>, <paramref name="title"/>, or <paramref name="description"/> is <see langword="null"/>.</exception>
-    public EmbeddedExternal(string uri, string title, string description, Blob? thumbnail, StrongReference[] associatedRefs) :
-        this(new External.Properties(new Uri(uri), title, description, thumbnail, associatedRefs))
     {
         ArgumentNullException.ThrowIfNull(uri);
         ArgumentNullException.ThrowIfNull(title);