+++ src/idunno.Bluesky/RichText/LinkFacetFeature.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Serialization;
 
 namespace idunno.Bluesky.RichText;
@@ -16,7 +17,8 @@ public sealed record LinkFacetFeature : FacetFeature
     /// <param name="uri">The <see cref="Uri"/> for the facet.</param>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/> is <see langword="null"/>.</exception>
     [JsonConstructor]
-    public LinkFacetFeature(Uri uri)
+    [SuppressMessage("Design", "CA1054:URI-like parameters should not be strings", Justification = "The official Bluesky can create facets with illegal URIs, so this parameter is a string to accommodate those cases.")]
+    public LinkFacetFeature(string uri)
     {
         ArgumentNullException.ThrowIfNull(uri);
 
@@ -28,5 +30,6 @@ public sealed record LinkFacetFeature : FacetFeature
     /// </summary>
     [JsonInclude]
     [JsonRequired]
-    public Uri Uri { get; init; }
+    [SuppressMessage("Design", "CA1056:URI-like properties should not be strings", Justification = "The official Bluesky can create facets with illegal URIs, so this property is a string to accommodate those cases.")]
+    public string Uri { get; init; }
 }